Language.Haskell.Tools.Daemon.ErrorHandling

Description

Handlers for common errors in Haskell-tools daemon.

Documentation

userExceptionHandlers :: (String -> IO a) -> ([Marker] -> [String] -> IO a) -> [Handler a]

exceptionHandlers :: IO () -> (String -> IO ()) -> [Handler ()]

Handlers for generic exceptions: IOException, AsyncException, SomeException.

getProblems :: SourceError -> ([Marker], [String])

handleGHCException :: String -> Maybe (String, Bool)

Hint text and continuation suggestion for different kinds of errors based on pattern matching on error text.

handleSourceProblem :: String -> Maybe String

Hint text and continuation suggestion for different kinds of source problems based on pattern matching on error text.