Language.Haskell.Tools.Daemon.Mode

Description

Defines different working modes for the daemon. It can work by using a socket connection or channels to communicate with the client. When the daemon is used by CLI, it uses channel if it communicates with an editor plugin it uses the socket connection.

Documentation

data WorkingMode a

An abstraction over the connection to the client.

Constructors

WorkingMode

Fields

daemonConnect :: Int -> IO a

daemonDisconnect :: a -> IO ()

daemonSend :: a -> ResponseMsg -> IO ()

daemonReceive :: a -> IO [Either String ClientMessage]

socketMode :: WorkingMode (Socket, Socket)

Connect to the client running in a separate process using socket connection

channelMode :: WorkingMode (Chan ResponseMsg, Chan ClientMessage)

Connect to the client running in the same process using a channel