Language.Haskell.Tools.Daemon.PackageDB

Description

Setting the package database to use when compiling modules. The daemon must have one single package database that cannot be changed after a package is loaded using that package database. Available package databases are the cabal global, the cabal sandbox, the stack or one that had been explicitely set by a file path.

Documentation

data PackageDB

Possible package database configurations.

Constructors

DefaultDB
Use the global cabal package database (like when using ghc).

CabalSandboxDB
Use the sandboxed cabal package database.

StackDB
Use the stack package databases (local and snapshot).

ExplicitDB
Set the package database explicitely.

Fields

packageDBPath :: [FilePath]

decidePkgDB :: [FilePath] -> IO (Maybe PackageDB)

Decide which type of project we are dealing with based on the package folders. Should only be invoked if the user did not select the project-type.

packageDBLoc :: PackageDB -> FilePath -> IO [FilePath]

Finds the location of the package database based on the configuration.

detectAutogen :: FilePath -> PackageDB -> IO (Maybe FilePath)

Gets the (probable) location of autogen folder depending on which type of build we are using.