Language.Haskell.Tools.Daemon.Protocol

Description

This module declares the messages that can be sent from the client to the daemon engine and from the engine to the client.

Documentation

data ClientMessage

The messages expected from the client.

Constructors

KeepAlive
A simple ping message to check that the server is running.

Reset
A message that instructs the server to reset its internal state and re-load loaded packages.

Handshake
Tells the client version and asks the servers version.

Fields

clientVersion :: [Int]

SetPackageDB
Sets the package database for the engine to use.

Fields

pkgDB :: PackageDB

AddPackages
Registers packages to the engine. They will be subject to subsequent refactorings. Will cause the packages to be loaded, resulting in LoadingModules, LoadedModule or CompilationProblem responses.

Fields

addedPathes :: [FilePath]

RemovePackages
Deregisters the given packages from the engine. They will not be subjects of further refactorings.

Fields

removedPathes :: [FilePath]

SetWorkingDir
Sets the working directory for the compilation. Important when compiling code that loads resources based on relative pathes.

Fields

newWorkingDir :: FilePath

SetGHCFlags
Sets the compilation flags. The unused flags are returned via the UnusedFlags response.

Fields

ghcFlags :: [String]

PerformRefactoring
Orders the engine to perform the refactoring on the module given with the selection and details. Successful refactorings will cause re-loading of modules. If shutdownAfter or diffMode is not set, after the refactoring, modules are re-loaded, LoadingModules, LoadedModule responses are sent.

Fields

refactoring :: String

modulePath :: FilePath

editorSelection :: String

details :: [String]

Additional details for the refactoring like the names of generated definitions.

shutdownAfter :: Bool

Stop the daemon after performing the refactoring.

diffMode :: Bool

Don't change the files, send back the result as a unified diff.

UndoLast
Asks the daemon to undo the last refactoring.

Disconnect
Stops the engine. It replies with Disconnected.

ReLoad
Instructs the engine to re-load a changed module. LoadingModules, LoadedModule responses may be sent.

Fields

addedModules :: [FilePath]

changedModules :: [FilePath]

removedModules :: [FilePath]

Stop
Stops the server. OBSOLATE

data ResponseMsg

The possible responses that the server can give.

Constructors

KeepAliveResponse
A response to KeepAlive

HandshakeResponse
Tells the version of the server.

Fields

serverVersion :: [Int]

ErrorMessage
An error message marking internal problems or user mistakes. TODO: separate internal problems and user mistakes.

Fields

errorMsg :: String

CompilationProblem
A response that tells there are errors in the source code given.

Fields

markers :: [Marker]

errorHints :: [String]

DiffInfo
Information about changes that would be caused by the refactoring.

Fields

diffInfo :: String

LoadingModules
The traversal of the project is done, now the engine is loading the given modules.

Fields

modulesToLoad :: [FilePath]

LoadedModule
The engine has loaded the given module.

Fields

loadedModulePath :: FilePath

loadedModuleName :: String

UnusedFlags
Returns the flags that are not used by the engine.

Fields

unusedFlags :: [String]

Disconnected
The engine has closed the connection.

data Marker

Constructors

Marker

Fields

location :: SrcSpan

severity :: Severity

message :: String

data Severity

Constructors

Error

Warning

Info

data UndoRefactor

Constructors

RemoveAdded

Fields

undoRemovePath :: FilePath

RestoreRemoved

Fields

undoRestorePath :: FilePath

undoRestoreContents :: String

UndoChanges

Fields

undoChangedPath :: FilePath

undoDiff :: FileDiff

type FileDiff = [(Int, Int, String)]

Orphan instances

ToJSON SrcSpan

Methods

toJSON :: SrcSpan -> Value

toEncoding :: SrcSpan -> Encoding

toJSONList :: [SrcSpan] -> Value

toEncodingList :: [SrcSpan] -> Encoding