Language.Haskell.Tools.Daemon.Representation

Description

Representation of the modules and packages in the daemon session.

Documentation

data ModuleCollection k

The modules of a library, executable, test or benchmark. A package contains one or more module collection.

Constructors

ModuleCollection

Fields

_mcId :: ModuleCollectionId

_mcLoadDone :: Bool

_mcRoot :: FilePath

_mcSourceDirs :: [FilePath]

_mcModuleFiles :: [(ModuleNameStr, FilePath)]

_mcModules :: Map k ModuleRecord

_mcFlagSetup :: DynFlags -> IO DynFlags

Sets up the ghc environment for compiling the modules of this collection

_mcLoadFlagSetup :: DynFlags -> IO DynFlags

Sets up the ghc environment for dependency analysis

_mcDependencies :: [ModuleCollectionId]

modCollToSfk :: ModuleCollection ModuleNameStr -> ModuleCollection SourceFileKey

data ModuleRecord

The state of a module.

Constructors

ModuleNotLoaded

Fields

_modRecCodeGen :: CodeGenPolicy

_recModuleExposed :: Bool

ModuleParsed

Fields

_parsedRecModule :: UnnamedModule

_modRecMS :: ModSummary

ModuleRenamed

Fields

_renamedRecModule :: UnnamedModule

_modRecMS :: ModSummary

ModuleTypeChecked

Fields

_typedRecModule :: UnnamedModule

_modRecMS :: ModSummary

_modRecCodeGen :: CodeGenPolicy

isLoaded :: ModuleRecord -> Bool

data CodeGenPolicy

Constructors

NoCodeGen

InterpretedCode

GeneratedCode

type ModuleNameStr = String

An alias for module names

data ModuleCollectionId

This data structure identifies a module collection.

Constructors

DirectoryMC FilePath

LibraryMC String

ExecutableMC String String

TestSuiteMC String String

BenchmarkMC String String

mcSourceDirs :: forall k. Lens (ModuleCollection k) (ModuleCollection k) [FilePath] [FilePath]

mcRoot :: forall k. Lens (ModuleCollection k) (ModuleCollection k) FilePath FilePath

mcModules :: forall k k'. Lens (ModuleCollection k) (ModuleCollection k') (Map k ModuleRecord) (Map k' ModuleRecord)

mcModuleFiles :: forall k. Lens (ModuleCollection k) (ModuleCollection k) [(ModuleNameStr, FilePath)] [(ModuleNameStr, FilePath)]

mcLoadFlagSetup :: forall k. Lens (ModuleCollection k) (ModuleCollection k) (DynFlags -> IO DynFlags) (DynFlags -> IO DynFlags)

mcLoadDone :: forall k. Lens (ModuleCollection k) (ModuleCollection k) Bool Bool

mcId :: forall k. Lens (ModuleCollection k) (ModuleCollection k) ModuleCollectionId ModuleCollectionId

mcFlagSetup :: forall k. Lens (ModuleCollection k) (ModuleCollection k) (DynFlags -> IO DynFlags) (DynFlags -> IO DynFlags)

mcDependencies :: forall k. Lens (ModuleCollection k) (ModuleCollection k) [ModuleCollectionId] [ModuleCollectionId]

typedRecModule :: Partial ModuleRecord ModuleRecord UnnamedModule UnnamedModule

renamedRecModule :: Partial ModuleRecord ModuleRecord UnnamedModule UnnamedModule

recModuleExposed :: Partial ModuleRecord ModuleRecord Bool Bool

parsedRecModule :: Partial ModuleRecord ModuleRecord UnnamedModule UnnamedModule

modRecMS :: Partial ModuleRecord ModuleRecord ModSummary ModSummary

modRecCodeGen :: Partial ModuleRecord ModuleRecord CodeGenPolicy CodeGenPolicy