Language.Haskell.Tools.Daemon.State

Documentation

data DaemonSessionState

The actual state of the daemon process. Contains loaded modules and user settings. The GHC state is handled separately.

Constructors

DaemonSessionState

Fields

_refSessMCs :: [ModuleCollection SourceFileKey]

The package components loaded into the session.

_packageDB :: Maybe (PackageDB, Bool)

The package database that is selected and a flag to decide if it is forced.

_ghcFlagsSet :: DynFlags -> DynFlags

GHC flags for compiling modules. Overrides settings in cabal files.

_pkgDbFlags :: DynFlags -> DynFlags

_packageDBLocs :: [FilePath]

The pathes where the package databases are located.

_exiting :: Bool

True if in the process of shutting down the session.

_undoStack :: [[UndoRefactor]]

True if in the process of shutting down the session.

_watchProc :: Maybe WatchProcess

Information about the file system watch process.

_watchThreads :: [ThreadId]

Extra threads started for handling the information from the watch process.

_touchedFiles :: Set FilePath

Marks files as being modified by this process Changes detected on marked files will not invalidate refactoring history.

initSession :: DaemonSessionState

An initial state of a daemon session.

resetSession :: DaemonSessionState -> DaemonSessionState

watchThreads :: Lens DaemonSessionState DaemonSessionState [ThreadId] [ThreadId]

watchProc :: Lens DaemonSessionState DaemonSessionState (Maybe WatchProcess) (Maybe WatchProcess)

undoStack :: Lens DaemonSessionState DaemonSessionState [[UndoRefactor]] [[UndoRefactor]]

touchedFiles :: Lens DaemonSessionState DaemonSessionState (Set FilePath) (Set FilePath)

refSessMCs :: Lens DaemonSessionState DaemonSessionState [ModuleCollection SourceFileKey] [ModuleCollection SourceFileKey]

pkgDbFlags :: Lens DaemonSessionState DaemonSessionState (DynFlags -> DynFlags) (DynFlags -> DynFlags)

packageDBLocs :: Lens DaemonSessionState DaemonSessionState [FilePath] [FilePath]

packageDB :: Lens DaemonSessionState DaemonSessionState (Maybe (PackageDB, Bool)) (Maybe (PackageDB, Bool))

ghcFlagsSet :: Lens DaemonSessionState DaemonSessionState (DynFlags -> DynFlags) (DynFlags -> DynFlags)

exiting :: Lens DaemonSessionState DaemonSessionState Bool Bool