Language.Haskell.Tools.Daemon.Utils

Description

Utility operations for the reprsentation of module collections.

Documentation

moduleCollectionPkgId :: ModuleCollectionId -> Maybe String

Get the name of a module collection from its Id.

lookupModuleColl :: String -> [ModuleCollection SourceFileKey] -> Maybe (ModuleCollection SourceFileKey)

Find the module collection where the given module is contained. Based on module name.

lookupSourceFileColl :: FilePath -> [ModuleCollection SourceFileKey] -> Maybe (ModuleCollection SourceFileKey)

Find the module collection where the given module is contained. Based on source file name.

lookupModuleInSCs :: String -> [ModuleCollection SourceFileKey] -> Maybe (SourceFileKey, ModuleRecord)

Find the module with the given name. Based on module name.

lookupModInSCs :: SourceFileKey -> [ModuleCollection SourceFileKey] -> Maybe (SourceFileKey, ModuleRecord)

Find the module with the given name. Based on source file name.

lookupSFKInSCs :: SourceFileKey -> [ModuleCollection SourceFileKey] -> Maybe ModuleRecord

Get the module with the given name and source file key.

removeModule :: String -> [ModuleCollection SourceFileKey] -> [ModuleCollection SourceFileKey]

Remove a module with the given name. Based on module name.

hasGeneratedCode :: SourceFileKey -> [ModuleCollection SourceFileKey] -> Bool

Check if we already generated code for the given module. Based on both module name and source file name.

needsGeneratedCode :: SourceFileKey -> [ModuleCollection SourceFileKey] -> Bool

Check if the given module needs code generation. Finds the module if no source file name is present and module names check or if both module names and source file names check.

codeGeneratedFor :: SourceFileKey -> [ModuleCollection SourceFileKey] -> [ModuleCollection SourceFileKey]

Marks the given module for code generation. Finds the module if no source file name is present and module names check or if both module names and source file names check.

isAlreadyLoaded :: SourceFileKey -> [ModuleCollection SourceFileKey] -> Bool

Check if the given module has been already loaded. Based on both module name and source file name.

insertIfMissing :: SourceFileKey -> [ModuleCollection SourceFileKey] -> [ModuleCollection SourceFileKey]

Insert a module with a source file key to our database if it wasn't there already