Language.Haskell.Tools.Daemon.Watch

Description

Controls the file system watching in the daemon. The file system watching must run in a separate process to prevent blocking because of file operations interfering with watch.

Documentation

createWatchProcess' :: Maybe FilePath -> Session -> MVar DaemonSessionState -> MVar [Marker] -> (ResponseMsg -> IO ()) -> IO (Maybe WatchProcess, [ThreadId])

Starts the watch process and a thread that receives notifications from it. The notification thread will invoke updates on the daemon state to re-load files.

stopWatch :: WatchProcess -> [ThreadId] -> IO ()

Stops the watch process and all threads associated with it.