Language.Haskell.Tools.Daemon

Description

The central module for the background process of Haskell-tools. Starts the daemon process and updates it for each client request in a loop. After this releases the resources and terminates.

Documentation

runDaemon' :: [RefactoringChoice IdDom] -> DaemonOptions -> IO ()

Starts the daemon process. This will not return until the daemon stops. You can use this entry point when the other endpoint of the client connection is not needed, for example, when you use socket connection to connect to the daemon process.

runDaemon :: [RefactoringChoice IdDom] -> WorkingMode a -> MVar a -> DaemonOptions -> IO ()

Starts the daemon process. This will not return until the daemon stops. The daemon process is parameterized by the refactorings you can use in it. This entry point gives back the other endpoint of the connection so it can be used to run the daemon in the same process.

serverLoop :: [RefactoringChoice IdDom] -> WorkingMode a -> a -> DaemonOptions -> Session -> MVar DaemonSessionState -> IO ()

Starts the server loop, receiving requests from the client and updated the server state according to these.

respondTo :: DaemonOptions -> [RefactoringChoice IdDom] -> Session -> MVar DaemonSessionState -> (ResponseMsg -> IO ()) -> ClientMessage -> IO Bool

Responds to a client request by modifying the daemon and GHC state accordingly.