Language.Haskell.Tools.Daemon.State

Documentation

data DaemonSessionState

The actual state of the daemon process. Contains loaded modules and user settings. The GHC state is handled separately.

Constructors

DaemonSessionState

Fields

_refSessMCs :: [ModuleCollection SourceFileKey]

The package components loaded into the session.

_packageDB :: PackageDB

The package database that is selected.

_ghcFlagsSet :: DynFlags -> DynFlags

GHC flags for compiling modules. Overrides settings in cabal files.

_pkgDbFlags :: DynFlags -> DynFlags

_packageDBSet :: Bool

True if the package database is actually used. The package database cannot be changed if set.

_packageDBLocs :: [FilePath]

The pathes where the package databases are located.

_exiting :: Bool

True if in the process of shutting down the session.

_undoStack :: [[UndoRefactor]]

True if in the process of shutting down the session.

_watchProc :: Maybe WatchProcess

Information about the file system watch process.

_watchThreads :: [ThreadId]

Extra threads started for handling the information from the watch process.

initSession :: DaemonSessionState

An initial state of a daemon session.

resetSession :: DaemonSessionState -> DaemonSessionState

watchThreads :: Lens DaemonSessionState DaemonSessionState [ThreadId] [ThreadId]

watchProc :: Lens DaemonSessionState DaemonSessionState (Maybe WatchProcess) (Maybe WatchProcess)

undoStack :: Lens DaemonSessionState DaemonSessionState [[UndoRefactor]] [[UndoRefactor]]

refSessMCs :: Lens DaemonSessionState DaemonSessionState [ModuleCollection SourceFileKey] [ModuleCollection SourceFileKey]

pkgDbFlags :: Lens DaemonSessionState DaemonSessionState (DynFlags -> DynFlags) (DynFlags -> DynFlags)

packageDBSet :: Lens DaemonSessionState DaemonSessionState Bool Bool

packageDBLocs :: Lens DaemonSessionState DaemonSessionState [FilePath] [FilePath]

packageDB :: Lens DaemonSessionState DaemonSessionState PackageDB PackageDB

ghcFlagsSet :: Lens DaemonSessionState DaemonSessionState (DynFlags -> DynFlags) (DynFlags -> DynFlags)

exiting :: Lens DaemonSessionState DaemonSessionState Bool Bool