Language.Haskell.Tools.Daemon.Update

Description

Resolves how the daemon should react to individual requests from the client.

Documentation

updateClient :: DaemonOptions -> MVar [Marker] -> [RefactoringChoice] -> (ResponseMsg -> IO ()) -> ClientMessage -> DaemonSession Bool

This function does the real job of acting upon client messages in a stateful environment of a client.

updateForFileChanges :: (ResponseMsg -> IO ()) -> MVar [Marker] -> [FilePath] -> [FilePath] -> [FilePath] -> DaemonSession ()

Updates the state of the tool after some files have been changed (possibly by another application)

initGhcSession :: Bool -> IO (Session, MVar [Marker])