Language.Haskell.Tools.Daemon.Update

Description

Resolves how the daemon should react to individual requests from the client.

Documentation

updateClient :: [RefactoringChoice IdDom] -> (ResponseMsg -> IO ()) -> ClientMessage -> DaemonSession Bool

This function does the real job of acting upon client messages in a stateful environment of a client.

addPackages :: (ResponseMsg -> IO ()) -> [FilePath] -> DaemonSession ()

reloadModules :: (ResponseMsg -> IO ()) -> [FilePath] -> [FilePath] -> [FilePath] -> DaemonSession Bool

Reloads changed modules to have an up-to-date version loaded

createUndo :: Eq a => Int -> [Diff [a]] -> [(Int, Int, [a])]

Creates a compressed set of changes in one file

createUnifiedDiff :: FilePath -> String -> String -> String

Creates a unified-style diff of two texts. Only used when the user wants to know what would change.

performUndo :: UndoRefactor -> IO ()

Undo a refactoring change using the information that was saved earlier.

performUndoChanges :: Int -> FileDiff -> String -> String

Undo the changes in one file using the information that was saved earlier. See createUndo.

getUndoAdded :: [UndoRefactor] -> [FilePath]

Get the files added by a refactoring.

getUndoChanged :: [UndoRefactor] -> [FilePath]

Get the files changed by a refactoring.

getUndoRemoved :: [UndoRefactor] -> [FilePath]

Get the files removed by a refactoring.

initGhcSession :: IO Session

usePackageDB :: GhcMonad m => [FilePath] -> m ()

watchNew :: FilePath -> DaemonSession ()